Website got updated!
Sorry for the longest wait. The site got a small upgrade. Did my best attempt to write and maintain the proper flow clearly as possible, I tried to re-articulated and adjusted some things, I do believe still needs more improvement but this is doable for now. More updates are planned in the near future. I am releasing this as-is for now.